Move configuration conversion to a separate single use class
